requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
redis==5.0.0
pydantic==2.5.3
gunicorn==21.2.0
pytest==7.4.4
python-dotenv==1.0.0
SQLAlchemy==2.0.20
alembic==1.13.1
bcrypt==4.0.1
flask-cors==3.0.10
